+++ UCR.Core/Models/Subscription/PluginSubscription.cs
@@ -10,7 +10,7 @@ namespace HidWizards.UCR.Core.Models.Subscription
         public Guid SubscriptionStateGuid { get; set; }
         public List<OutputSubscription> OutputSubscriptions { get; set; }
 
-        public PluginSubscription(Plugin plugin, Guid subscriptionStateGuid, List<DeviceSubscription> outputDeviceSubscriptions)
+        public PluginSubscription(Mapping mapping, Plugin plugin, Guid subscriptionStateGuid, List<DeviceConfigurationSubscription> outputDeviceConfigurations)
         {
             Plugin = plugin;
             SubscriptionStateGuid = subscriptionStateGuid;
@@ -19,8 +19,15 @@ namespace HidWizards.UCR.Core.Models.Subscription
             foreach (var deviceBinding in Plugin.Outputs)
             {
                 if (!deviceBinding.IsBound) continue;
-                var outputDeviceSubscription = outputDeviceSubscriptions.FirstOrDefault(d => d.Device.Guid == deviceBinding.DeviceGuid);
-                OutputSubscriptions.Add(new OutputSubscription(deviceBinding, subscriptionStateGuid, outputDeviceSubscription));
+                
+                var deviceConfigurationSubscription = outputDeviceConfigurations.FirstOrDefault(configuration => configuration.DeviceConfiguration.Guid == deviceBinding.DeviceConfigurationGuid);
+                if (deviceConfigurationSubscription == null) continue;
+
+                var deviceConfiguration = mapping.IsShadowMapping
+                    ? deviceConfigurationSubscription.ShadowDeviceSubscriptions[mapping.ShadowDeviceNumber]
+                    : deviceConfigurationSubscription.DeviceSubscription;
+
+                OutputSubscriptions.Add(new OutputSubscription(deviceBinding, subscriptionStateGuid, deviceConfiguration));
             }
         }
     }