+++ UCR.Core/Models/Subscription/SubscriptionState.cs
@@ -9,7 +9,7 @@ namespace HidWizards.UCR.Core.Models.Subscription
         public Profile ActiveProfile { get; }
         public bool IsActive { get; set; }
 
-        public List<DeviceSubscription> OutputDeviceSubscriptions { get; }
+        public List<DeviceConfigurationSubscription> OutputDeviceConfigurationSubscriptions { get; }
         public List<MappingSubscription> MappingSubscriptions { get; set; }
 
 
@@ -17,19 +17,18 @@ namespace HidWizards.UCR.Core.Models.Subscription
         {
             StateGuid = Guid.NewGuid();
             ActiveProfile = profile;
-            OutputDeviceSubscriptions = new List<DeviceSubscription>();
+            OutputDeviceConfigurationSubscriptions = new List<DeviceConfigurationSubscription>();
             MappingSubscriptions = new List<MappingSubscription>();
             IsActive = false;
         }
 
-        public DeviceSubscription AddOutputDevice(Device device, Profile profile)
+        public void AddOutputDeviceConfiguration(DeviceConfiguration deviceConfiguration)
         {
-            var deviceSubscription = new DeviceSubscription(device, profile);
-            OutputDeviceSubscriptions.Add(deviceSubscription);
-            return deviceSubscription;
+            var deviceSubscription = new DeviceConfigurationSubscription(deviceConfiguration);
+            OutputDeviceConfigurationSubscriptions.Add(deviceSubscription);
         }
         
-        public void AddMappings(Profile profile, List<DeviceSubscription> profileOutputDevices)
+        public void AddMappings(Profile profile, List<DeviceConfigurationSubscription> profileOutputDevices)
         {
             var profileMappings = new List<MappingSubscription>();
 
@@ -39,7 +38,37 @@ namespace HidWizards.UCR.Core.Models.Subscription
             }
 
             OverrideParentMappings(profileMappings);
+
             MappingSubscriptions.AddRange(profileMappings);
+            MappingSubscriptions.AddRange(AddShadowMappings(profile, profileMappings, profileOutputDevices));
+        }
+
+        private List<MappingSubscription> AddShadowMappings(Profile profile, List<MappingSubscription> profileMappings, List<DeviceConfigurationSubscription> profileOutputDevices)
+        {
+            var result = new List<MappingSubscription>();
+
+            foreach (var mappingSubscription in profileMappings)
+            {
+                var shadowClones = mappingSubscription.Mapping.PossibleShadowClones;
+                if (shadowClones == 0) continue;
+                
+
+                result.AddRange(CloneMappingSubscription(profile, mappingSubscription, profileOutputDevices, shadowClones));
+            }
+
+            return result;
+        }
+
+        private List<MappingSubscription> CloneMappingSubscription(Profile profile, MappingSubscription mappingSubscription, List<DeviceConfigurationSubscription> profileOutputDevices, int shadowClones)
+        {
+            var result = new List<MappingSubscription>();
+
+            for (var i = 0; i < shadowClones; i++)
+            {
+                result.Add(new MappingSubscription(profile, mappingSubscription.Mapping.CreateShadowClone(i), StateGuid, profileOutputDevices));
+            }
+
+            return result;
         }
 
         private void OverrideParentMappings(List<MappingSubscription> profileMappingSubscriptions)